requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.1
Django==3.2.12
django-cors-headers==3.11.0
django-phonenumber-field==6.1.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
phonenumbers==8.12.43
psycopg2==2.9.3
pycparser==2.21
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
